move a `snapshot/add/` test into `testsuite/cargo_add` directory
