+++ tests/testsuite/cargo_add.rs
@@ -2,6 +2,8 @@ use cargo_test_support::compare::assert;
 use cargo_test_support::prelude::*;
 use cargo_test_support::Project;
 
+mod build_prefer_existing_version;
+
 fn init_registry() {
     cargo_test_support::registry::init();
     add_registry_packages(false);
@@ -200,28 +202,6 @@ fn build() {
         .stderr_matches_path("tests/snapshots/add/build/stderr.log");
 
     assert().subset_matches("tests/snapshots/add/build/out", &project_root);
-}
-
-#[cargo_test]
-fn build_prefer_existing_version() {
-    init_alt_registry();
-    let project = Project::from_template("tests/snapshots/add/build_prefer_existing_version/in");
-    let project_root = project.root();
-    let cwd = &project_root;
-
-    snapbox::cmd::Command::cargo()
-        .arg("add")
-        .arg_line("cargo-list-test-fixture-dependency --build")
-        .current_dir(cwd)
-        .assert()
-        .success()
-        .stdout_matches_path("tests/snapshots/add/build_prefer_existing_version/stdout.log")
-        .stderr_matches_path("tests/snapshots/add/build_prefer_existing_version/stderr.log");
-
-    assert().subset_matches(
-        "tests/snapshots/add/build_prefer_existing_version/out",
-        &project_root,
-    );
 }
 
 #[cargo_test]
