+++ tests/testsuite/cargo_add/build_prefer_existing_version/in/dependency/Cargo.toml
